app/requirements.txt
anyio==3.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
db-dtypes==1.0.5
fastapi==0.92.0
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.5.0
google-cloud-bigquery-storage==2.18.1
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pandas-gbq==0.19.1
proto-plus==1.22.2
protobuf==4.22.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
pydata-google-auth==1.7.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
